simple implementation of summon passives
uses combination of passive=true and buff_party=true with optional
buff_party_power_id. works on summons only and not converted enemies.
also the passives applied are the passives available at the time of
summoning i.e not recalculated for existing summons. also added back
some code which was incorrectly removed in the last commit.