Add --game command line flag
This allows for the creation/search of directories under a single
"flare" directory. For example, if we launch Flare with `flare
--game=wandercall`: config files will be in
`$CONFIG_DIR/flare/wandercall/`; saves will be in
`$LOCAL_DATA_DIR/flare/wandercall/`; and game data (mods folder) will be
in `$DATA_DIR/flare/wandercall/`.

If the flag is not set, we fall back to the current directory.