Fix teleporting to spawn map when a map is missing
Previously when a map was missing, the player would be teleported to the
spawn map. They'd be stuck there because the teleportation flag had been
set to false, since they already teleported once. This commit allows for
a sort of "chain teleporting", where we can keep doing intermap
teleports until reaching an endpoint.