Rename class EnemyBehavior to EntityBehavior
The class BehaviorAlly is inheriting from BehaviorStandard, which itself
inherited from EnemyBehavior before.
As allies are no enemies, we should go with the broader term
EntityBehavior.

Technically speaking this abstract base class just works with enemy objects
and not on the abstract entity objects, which is missleading on the
technical side.