Introduce new Map class
The new class Map deals with loading a map.
The game related stuff, such as rendering, dealing with events and such
is left to the MapRenderer class, which inherits from the Map class.

This will later on be interesting for randomly generated maps as we need
to load some kind of Automapping maps to nice up the schematic.