Use chance_flee for enemies
Closes #103

This allows for improved ranged enemy behavior. A roll of chance_flee
will determine if the enemy will attempt to move to a safe distance
(threat_range/2). Once the enemy has reached that distance, they will
not pursue the hero. When the hero gets close to the enemy, it will
either try moving to a safe distance again, or it will pursue the hero.