Add support for mod dependencies
Closes #733

In mods/$MOD/settings.txt, use:

requires=mod1,mod2,etc

If there's a missing dependency, the game will search all available mods
for it. If it still can't find it, the mod that requires the missing mod
will be disabled. applyDepends() will run recursivly until all
dependencies are met and in the correct order.