MapRenderer: remove caching of background image for animated tiles off
When animated tiles is set to 'off', we currently had an optimization,
which merged all the background layers into one large image, which needed
to be recalculated only occasionally instead of every frame.

This had a few disadvantages:
* When it needed to be regenerated, it took longer. So the game was lagging
  every few seconds instead of just being slow.
* The memory needed for the animated tiles being off was larger than the
  memory needed with animated tiles.

By removing that optimization, we should use the same amount of memory
instead of more (fixes #908) and this should also address the occasional
glitches, which occur currently in the SDL2 port (fixes #928).