MenuInventory: Fix potential memory access fault
When dropping vendor items on to a filled inventory slot, 'drag_prev_src'
is -1, so 'drag_prev_slot' will be calculated to be
inventory[-1].drag_prev_slot. However that value is never used.

As we don't know which memory we're accessing with inventory[-1]
(Well, some random memory before the inventory definition, which
currently luckily happens to be still member variables of MenuInventory),
we may get an memory access violation. As we're not using the computed
value, we're safe to omit calculating it in the first place.

Also see fe2dfbd18693f317 (2012-10-02, Fix crash when dropping vendor
items on to filled inventory slots)