Simplify checkNPCInteraction and improve MenuManager's closeAll()
A lot of the code in GameStatePlay::checkNPCInteraction() was from when
I added a Trader button to MenuTalker. That button has since been
replaced with the more flexible MenuNPCActions, so the code is not
needed anymore.

As for MenuManager, I changed closeAll() (including closeLeft() and
closeRight()) to work even when we're dragging something. Now, the
behavior is to return whatever we were dragging and *then* close the
menu(s). This should prevent dragging things when they shouldn't be
dragged.