Complete attribute documentaion
This should cover all of Flare's moddable configuration files. Some of
the descriptions might not be perfect, but at least we'll have an
overview of everything we can put in these files.

Of course, the Attribute Documentation wiki page should be regenerated
and updated after this. I added a few new types that should be added to
the top of that page:

* float - a floating point number
* alignment - can be any of: topleft, top, topright, left, center, right, bottomleft, bottom, bottomright
* label - specified as: "hidden" or as: x (integer), y (integer), justify (string), valign (string), style (string); justify can be [left:center:right]; valign can be [top:center:bottom]; style is any font style defined in engine/font_settings.txt

Known issue: The only functionality change I made was to the position
attribute in menus/fps.txt. I changed the old parsing of the corner to
place it in to use the alignment type mentioned above. Yay consistency!