Implement more consistent file parsing method
Closes #1040

Now by default, all mod files will overwrite eachother. For files that
were automatically concatenated before, they can still be concatenated
by making the first line of all but the first file "APPEND". For
example, we can combine the credits for the engine, fantasycore, and
alpha_demo by adding "APPEND" to the top of the credits cutscene files
in fantasycore and alpha_demo.