CMakeLists: make compatible with clang3.4
clang 3.4 is now standard at the travis building service and clang 3.4
seems to not support -fwhole-program.

Also remove the additional linking to lstdc++ if the compiler is clang.
We're a C++ project, so we'd want to use clang++ instead of clang.

Also removed -fno-enforce-eh-specs, which violates the C++ standard,
but may be useful for reducing code size in production builds.