Create a new Image when using resize()
This fixes a bug where cached images were being resized. The place
I noticed this was when turning on "scale_gfx" in cutscenes/credits.txt.
Since we use the Flare logo on both the title screen and the default
credits, the image was resized on the title screen after viewing the
credits.