Smoother movement around 1x1 objects/entities
When automatically maneuvering around 1x1 tiles, we were moving the
player to the outer edge of the tile instanly. This was jarring when the
size of the player's steps were smaller than that distance. This patch
makes it so that we approach that outer edge, but maintain our current
speed. The end result is much smoother movment around 1x1 objects.