Prevent log menu spam
This is useful for the "Game Saved" message we display now. If the game
gets quickly saved (like when organizing the stash), we now only display
the log message once. In the case of the HUD log, we reset the timer for
the last message, to make it appear like the message was triggered
again.