Allow items to disable equipment slots
This is useful for items like bows/slingshots, as they're essentially
"two-handed" weapons. So we can set "disable_slots=main" for bows to
prevent equipping stuff on the main hand.

Another use case for this could be a pair of overalls, which would go
into the chest slot, but "occupy" chest and legs.

It possible to disable more than one slot by specifying a list of slots
as opposed to just a single type.