Remove power_mod; Allow items to replace powers on the action bar
If an item has the line:

    replace_power=1,2

Then any instance of power 1 will be replaced with power 2 in the action
bar. Also it is possible to use:

    meta_power=true

for power definitions. The meta_power variable makes it so that a power
can't be cast (meaning it should be replaced with an item).