Don't use pixel-precision for event hotspots
This was originally implemented as a solution to a bug with event
hotspots not being accurate (I believe flare-game's "stash" chest
brought it to light). However, flare-engine-next's SDL 2 hardware
renderer does not have an implementation of RenderDevice::checkPixel().
flare-engine-next still functions properly, so I believe that checking
for pixel-precision was a solution to a red herring. It can be safely
removed.

It's also worth mentioning that having pixel-precision for hotspots
actually made it *more* difficult to click on them in some cases. For
example, the floor grates that trigger spike traps in the Ancient Temple
map have a lot of transparent pixels, so clicks would go "through" the
grates.