Don't use press/hold to simulate MAIN2; Use only MAIN1 instead
This makes it so that actionbar powers are triggered immediately when
pressed with touchscreen controls. This is more appropriate for how
action RPGs are played. There are a few downsides though:

- Can't rearrange/remove actionbar icons once they're placed. They can
still be overwritten by dragging in new powers/items from other menus.
- Items can't be activated directly from the players inventory.
Consumables should be dragged to the actionbar and used that way.

In the process, I removed a lot of the __ANDROID__ ifdefs, since there
are non-Android devices that use touch controls (laptops with
touchscreens come to mind). Also, having lots of ifdefs is just plain
ugly. I'd like to keep their usage to a minimum.