Perform targeting in MenuActionBar::checkAction()
This allows us to use a different target depending on how we trigger the
power. If we trigger a power by clicking on its icon or by pressing
ACTIONBAR_USE, the target will be in front of where the player is facing
(at melee range). Otherwise, we simply use the mouse cursor's position.

This is a partial solution to #1168