Add drop_max and drop_radius options to engine/loot.txt
- drop_max specifies the maximum number of rolls that occur when we drop
loot. If we have a drop_max of 4, this means we can potentially drop up
to 4 pieces of random loot in a single loot event. "Fixed" loot is
unaffected by any of this.
-drop_radius is the distance (in tiles) from the point of the loot event
in which loot can drop. By default this is 1, but as drop_max is
increased, it may make sense to increase this value.