Add two-click process when clicking on actionbar icons
Example uses in flare-game are Burn and Teleport. Right clicking on one
of these powers in the actionbar will disable the rest of the actionbar.
We can cancel this be pressing the key bound to CANCEL, or we can use
the power that is queued up by MAIN1 clicking in the playing field.

Closes #1168