For touchscreen devices, first tap will show tooltip for WidgetSlot
Closes #1181

When using a touchscreen, tapping on a WidgetSlot (inventory, stash,
vendor, or power menu) will highlight it in the same fashion as when
using NO_MOUSE. As a result, the tooltip that matches that slot's
contents will be shown. Tapping again will dismiss the tooltip, allowing
regular interaction with the WidgetSlot.