Don't pause when restoring Android app, but do prevent frameskip
The previous method of handling this involved pausing the game by
bringing up the pause menu. Unfortunately, this only worked for
GameStatePlay, since the other game states don't have pause menus.

This commit should be a better solution. Right before checking if the
game is paused, we now check if the game was minimized and is returning
to the foreground (minimized AND restored flags set). Here, we do the
same thing that is done when the game is paused: update the delta time
to prevent skipping frames.