Remove GameStateResolution, add fallbacks to createContext()
When the confirmation dialog in GameStateResolution was displayed, it
was possible to resize the window, effectively changing the settings the
player would be confirming.

Now that the engine is not changing video modes when changing the window
size, there is less need for a confirmation dialog.

However, there are still some settings that recreate the rendering
context (fullscreen, hardware surfaces, etc), so it's important that
there's a way to fall back to previous settings if one of these causes
failure. That functionality has been moved to the RenderDevice itself
(SDLSoftwareRenderDevice only in this commit).