Fix only the base stage of upgradable passive powers being activated
Instead of scanning power_cell, which only contains the base powers, we
should scan power_cell_unlocked, which contains the base powers AND all
unlocked upgrades.

Because of that, it also means that all levels of a passive power will
be stacked. So if the base power adds 50 HP, and level 2 adds 50 HP, the
player will have an additional 100 HP when level 2 is unlocked. For this
reason, it is recommended to keep animation/icon effects on the base
power only.