Migrate Android project to Android Studio and Gradle
The ADT plugin for Eclipse no longer works with the latest version of
Eclipse. Meanwhile, Google has halted development of the plugin in favor
of Android Studio.

Switching over is a welcome change. Initially setting up Android Studio
is much easier since the NDK can be downloaded from its pacakage
manager. The Eclipse ADT plugin only had the SDKs and tools packages.

Another welcome change is that the environment paths for the SDK and NDK
are not part of a project file that would be checked into git. So I no
longer need to maintain my own branch with paths specific to my system.