Add "loot_count" property to enemies and events
Using the loot_count property, we can define the minimum and maximum
amount of loot drops that can come from an enemy or an event. This is
handy for giving the player a larger reward when defeating bosses or
opening special chests. The property is used as so:

loot_count=min_drops,max_drops

max_drops may be ommited, in which case min_drops is used.