Allow picking quantity when shift-clicking on items
The old behavior was to pickup/move only 1 item in a stack. Now, we use
MenuNumPicker to let the player select between 1 and the stack size.

This will be very useful for gamepad & touch input, which share the
shift-click functionality for moving items. There was previously no way
to move entire item stacks with these input methods.