Use "sticky dragging" only when needed
"Sticky dragging" is when the player starts to drag an icon and releases
the mouse button, yet the dragged icon "sticks" to the cursor.

This patch does two things:

- Sticky dragging is only enabled when using MenuNumPicker to select an
item quantity for dragging.
- Since we can't change the cursor position without "clicking" on
a touchscreen, the dragged icon is drawn in a fixed position when sticky
dragging on a touchscreen.