Add docs/ directory for local documentation
This provides a framework for generating simple HTML pages for offline
reference. To regenerate all pages, run `docs/gen/gen_all.sh`.

Currently, the only document is the Attribute Reference, which is
essentially the same as the online wiki page. I think this page is
valuable enough to provide as an offline version despite that.