Fix stats not being updated when dropping equipment on actionbar
Equipment items with powers (like the Aether Walkers) can be placed on
the actionbar so that said powers can be used. Dragging the item causes
it to be unequipped, but it is re-equipped when it is dropped on the
action bar. The problem was that we weren't recalculating the player's
stats when the item was equipped again.