Add "state_duration" property to Powers
In a nutshell, this lets us pause on the last frame of an attack
animation. This is done by setting state_duration to a length longer
than the state animation. So if our "cast" animation is 400ms, setting
"state_duration=600ms" will hold on the last frame for an additional
200ms. Times that are less than the state animation are simply ignored,
as we don't want to break triggering of active frames.

The intended application of this property is for a "shield bash" power.
Our block animation is only 66ms, and we want to have the player hold
their shield up for the duration of the charge.