Add "relative_pos" property to Powers
This enables us to have hazards follow the caster as they move around.
An example use case might be an area-of-effect power that damages
enemies that get near the player. It can also be used for "charge"
powers, where the caster moves forward in a straight line during the
attack.

Note: This does NOT take into account any changes in the caster's
direction after the hazard has been created. Which is acceptable for the
above given examples, but might appear odd in other use cases.