Add "charge_speed" property to Powers
With this property, we can push the caster in the direction they are
facing until their state animation finishes. By using a negative speed,
we can also push the caster back 180 degrees from the direction they are
facing.

The intended use of this property is for "charge" powers, like a shield
bash, an ariel swoop, or a spear thrust. Since negative speeds push the
caster back, I can imagine a game with guns where the guns have recoil.