Allow binding joystick axes to all actions, not just movement
Bindings that use joystick axes are saved in the same way as mouse
buttons. e.g. We store them as negative integers less than -1.

In addition to allowing joystick axes to be bound, this commit moves
some of the SDL functions that were in main.cpp and GameStateConfig to
the SDLInputState class.