Try to move enemy pos when in melee range, but line-of-sight fails
When an enemy gets close to the player, they stop moving and should
ideally attack the player. However, this wasn't happening when enemies
were around a corner from the player. The enemy's line-of-sight to the
player is required to attack, but it is blocked in this situation.

This patch makes it so that the enemy will continue moving towards the
player when the line-of-sight check fails. That should allow the enemy
to relocate to a better attacking position.

Thanks Pawonashi for catching this bug.