Improve targeting animation position
The changes here make the targeting animation position change for powers
that have a limited targeting range, such as melee powers. Since the
point of the targeting animation is to visually show where a power's
hazard starts, this is much better than the previous implementation
which just used the mouse cursor position.