Don't allow repeaters to hit the same enemy multiple times
This is a solution for the (old) issue outlined here:
https://github.com/clintbellanger/flare/issues/217

Here's how it works: A repeater will spawn one parent hazard and several
children under that parent. The parent is responsible for tracking which
entities have been hit by it or any of its children. In other words, if
a child hazard hits an enemy, it gets added to the *parent's* list of
collided entities.

When a parent hazard expires, we make the next available child the new
parent. When a child expires, we remove it from the parent's list of
children.