Improve movement of fleeing entities
Closes #1454

Previously, entities would only flee in the exact opposite direction of
the vector from the entity to their target. If this direction happened
to be perpendicular to a wall, the entity would get stuck on said wall.

This patch makes it so that entities will flee in a random direction in
a half-circle away from their target. In other words, they can go
straight (as before), 90 degrees left, 90 degrees right, or a direction
in between.