Prevent spamming windowResize() calls when resizing window
This commit makes it so that there is a 1/4 second delay after resizing
the window before windowResize() is called. The windowResize() call can
be a bit expensive (deleting and allocating a large surface), so we
don't want to run it every frame that the user is resizing the window.

Also, we now check for SDL_WINDOWEVENT_SIZE_CHANGED instead of
SDL_WINDOWEVENT_RESIZED. The SDL docs describe the former as being
triggered through system and API calls that resize the window. It is
also triggered when SDL_WINDOWEVENT_RESIZED is triggered, so we only
need to use SDL_WINDOWEVENT_SIZE_CHANGED.