When playing item sounds, restrict to one playback per SoundID
To better illustrate why this change was made, here is an example:

The side-quest I am working on has an NPC take a series of notes the
player has found at the end. Each note is represented as an item, and
uses the same "paper" sound effect. When the NPC takes all the notes,
each one of them tries to play their sound effect. Without this commit,
all of them will play at once, resulting in a very loud noise. The fix
is to create virtual channels based on the SoundID variable.