Add a way for touchscreen users to read carried books
How this is intended to work: When an item stack with a quantity of 1 is
selected (e.g. tooltip is shown), the player can tap on it to activate
it. "Tap" in this case results in picking up the item and placing it
back where it was within 1/3 of a second.

See the code comments as to why we're not currently using this for
activating non-book items.