Allow maps to set a default spawn position for the player
Previously, there was the "location" key for map headers that set
a default position that was *only applied when the player was
initialized*. The "hero_pos" key replaces this, and it applies to any
time the player changes maps.

The "location" key also defined a default starting direction. This has
been omitted in the "hero_pos" key since we should always be using the
player's current direction when transitioning maps.

As a guideline, all maps should have hero_pos set. A console warning
will be displayed if it is not set, in addition to defaulting to (0,0).