Add "static" event activation type
Static events are executed every frame, with the need of interaction
from the player. Previously, the only way to implement this
functionality was to create an "on_trigger" with no hotspot defined over
the entire map. Doing that made working with other map objects in Tiled
difficult. This should be a cleaner solution.