Allow misses to instead deal a reduced amount of damage
By default, the damage percentage is 0, which retains the existing
behavior of not hitting the target and displaying the "miss" combat
text. Should the percentage be greater than 0, it will determine the
damage done to the target. The combat text for the damage will be
displayed in the same style as the "miss" text.