Add engine options for setting crit/overhit damage multipliers
Overhits are a new feature where excessive accuracy can produce an
alternate form of critical hit. As an example, let's say our hero has
150% accuracy and our target has 40% avoidance. This gives a "true"
accuracy value of 110%. To get our chance of scoring an overhit, we
subtract 100 from that value, meaning there is a 10% to score an overhit
in this case.