Add text_trim_boundary to GameStateLoad config
The implementation of text shortening for game slots in the previous
commit was wrong. It did not account for when the character preview was
on the left side of the text. It also assumed that the character preview
images themselves would have some padding. There was also a copy-paste
bug that caused only the map name string to be affected.