Directories should cause fileExists() to fail
A noted example of a bug caused by directories being passed to
fileExists() occured when giving an empty filename to an intermap Event.
Since ModManager::locate() would supply part of the path, the engine
would crash trying to load a map that didn't exist. There are probably
other cases like this with resource loading that this patch should fix.