Add replace_by_effect property to Powers
Note, this does NOT work like the replace_power property of items, which
updates the powers on the action bar. The power replacement done by
replace_by_effect takes place when the power is cast. This means it is
ideal (although not required) for the new power to have identical
MP/status/etc requirements as the original power.