Preload Loot, Power, Effect animations to prevent freezing
It was the case that these animations would not be loaded, with
AnimationSet::load(), until their first appearance in a game session.
Reading from disk is slow! We don't want to do it in the middle of
gameplay if we can avoid it.

Of course, this means greater, but more consistent, RAM usage.