Enemy waypoint improvements
This fixes a few things with how enemy waypoints worked:

- Turn delay is effectively disabled upon reaching a waypoint. Doing so
prevents the enemy from moving in their current direction before moving
towards the next waypoint
- We now try to get the enemy as close as possible to the actual
waypoint position (center of tile)
- There were cases where an enemy could never reach their destination
point, and would loop back-and-forth on a tile. This has been solved by
testing the distance of the next move. If the distance is worse than the
current distance to the destination, we simply place the enemy at the
destination. This only occurs if the enemy is less than a half tile away
from their destination.