Set default pathfinding limit to 10% of the map size
I settled on this value after testing with the summoned zombie on Salted
Field (a 60x60 map). The zombie would get stuck when trying to approach
the goblin spearman because it required a limit of approx. 240. This was
much larger than the previous default of 60. It just so happens that 10%
of the map size is 360, which is fairly close to 240 with some breathing
room.

Version 0.90.06 -> 0.90.07