Attempt to fix pathfinding turn delay bug
In response to this commit: https://github.com/clintbellanger/flare-game/commit/c6c764d65ea052a9acb74871c20ae58054eedd90

See the comment added in BehaviorStandard.cpp for a description of
what's going on here. In summary, the solution is to speed up the turn
delay to match the enemy's movement speed.