Force quest items into player's inventory, dropping non-quest items
Before, if the player received a quest item while having a full
inventory, the quest item would be dropped on the ground. This could
cause a lot of problems if the game developer assumes the player will
obtain the item at that moment.

To rememdy this, we now try to drop non-quest items to make room for the
quest item.