Have engine set NPC collision tile, use BLOCKS_ENTITIES
Since it was easy for mappers to forget to add a collision tile for each
NPC, we'll just make having the engine set the collision tile the
default behavior.

Also, we now use BLOCKS_ENTITIES instead of BLOCKS_MOVEMENT_HIDDEN. The
latter would allow flying creatures to move through NPCs.