Allow classes to specify a set of hero options they can use
Previously, all hero options (i.e. portraits) were available to every
class. Now, modders can define a set of hero options that a class can
use. This is useful for distinct character races.

Note that this required a small change to the engine/hero_options.txt
syntax. An integer index must be defined for each option. Temporarily,
we'll remain forwards compatible and automatically set the index based
on order in the file. However, to keep the code clean, this forwards
compatibility will be removed before release.