Reorganize path settings on Android
The purpose of doing this is to put things in a more accessible location
for users. They can now access the config directory to view the log or
change an internal setting if neccessary.

Also, we prefer the internal SD card over the external SD card. This is
because writting to external SD cards with 3rd party file managers (like
the ones on F-Droid) can be impossible on Android >= 6.0.

PATH_CONF
1. INTERNAL_SD_CARD/Flare
2. EXTERNAL_SD_CARD/Flare
3. App internal storage (/data/...)

PATH_DATA
1. App external storage (usually internal sd card)
2. INTERNAL_SD_CARD/Flare

PATH_USER
1. INTERNAL_SD_CARD/Flare
2. EXTERNAL_SD_CARD/Flare