Add vendor_ratio_buyback to engine/loot.txt
This variable is used to control the price of items in the buyback after
the player leaves the map. It defaults to 0, which tells the engine to
just use vendor_ratio (or the item's price_sell), which was the previous
behavior for all buyback items. As an example, setting
vendor_ratio_buyback to 100 will make old items in the buyback cost
their normal buy price.