Fix crash when teleporting to map hero_pos with a summoned creature
This crash was happening because teleporting to the default spawn
position on a map uses position (-1,-1). Since relocating summoned
creatures uses get_random_neighbor() with a radius of 1, the only
possible position would be (0,0). The position (0,0) is blocked on all
of flare-game's maps, so the summoned creatures would try to use the
player's current position, which at this point is still (-1,-1).

So the fix is to simply update the player's position before running any
of the handleNewMap() rountines.