Initial fix for #1279
How it works: As we iterate over a horizontal row of tiles (south-west
to north-east), we now render the objects located on the previous tile
after rendering a tile. This fixes the majority of rendering, but leaves
two spots where the order can be wrong depending on the surrounding
tiles. These spots are to the south and the east of the object. Most of
the code in this commit handles those edge cases.

Notes:
- If there are short tiles that are not packed, objects can disappear
behind their bounding box. Possible solutions include checking for a
transparent pixel in the tiles, or cropping the object so that it does not
overlap the tile.
- This has not been tested with unpacked objects, or objects that are
larger than 1 tile. The results should still be better than before, it's
just that our edge-case handling might fail.