Remove MenuDevHUD
Since the dev console can now be used to inspect the cursor/player
positions, there's no need for this menu.

However, the DEV_HUD setting remains, as we currently use it to toggle
the drawing of hazard positions. Drawing other "invisible" elements in
this manner should be added in the near future.