Revert "MessageEngine: Create a single get() function using va_list"
Using va_list wasn't a good idea where we were overloading the usage of
%d for various integer types. If the size of int was not the same as
int64_t (such as on a Raspberry Pi), we would crash if there were
multiple arguments.

For the get() that expects "%d %s", we were using it for strings that
had "%s %d" while swapping the arguments. I've made a separate function
that keeps the arguments in the correct order.

This reverts commit 409cf975fdfe0ad120c373eb03b901c895104b35.