Refactor MenuPowers
This commit simplifies the how the data for MenuPowers is stored and
used. Base powers and upgrades are now link together, so handling
upgrades (and potential downgrades) is now much easier. Each cell, or
"group", contains its position data, a list of powers, and an index for
the currently enabled power.

Also added was a way to skip over invisible slots when using keyboard
navigation. Previously, these slots could be selected, but there would
be no indication as they weren't being rendered.