Fix regression in 5718d021 that caused missing text
The old fix failed on strings such as "Hot. Hot. Hot." (from
flare-game). This is because the contents of each token were the same,
so the loop exited early. Instead of checking the token contents, we
should instead check the token cursor position. If the cursor is the
same after getting a new token, it's a clear sign that getting a new
token failed. It's also nice this is faster than doing the string
compare.