Fixed shrinking tab widgets when widow was resized
Instead of explicitly calling calc_size() for tab text, we simpy get it
from the label bounds (which internally calls calc_size()). So as
a bonus, this should be faster as well.

I've also cleaned things up a bit by putting the contents of
updateHeader() directly in setMainArea() and removing the former.