Add Timer class to Utils
A common paradigm in our codebase is things that operate on a "timer",
as in they wait X amount of frames before doing something. This ended up
causing some inconsistencies where we'd increment a "tick" counter in
some cases, yet decrement a counter in other cases. in reality, such
timers can be abstracted as having a beginning, an end, and a way to
increment them.

For the most common use case, the following methods will likely be used:

- setDuration(unsigned): sets the timer duration in frames and resets to
the beginning
- tick(): increments the timer by 1
- isEnd(): the timer has reached the END state
- reset(int): resets the timer to either the BEGIN or END, depending on
the passed enum flag