Remove MenuNPCActions; show actions in MenuTalker
This should be the first step in making branching dialog choices more
feasible. Topics (including "Trade") are displayed in MenuTalker as
text-only buttons. WidgetButton has been extended to support this
functionality. Currently, topics and normal dialog text can't be
combined, but now it shouldn't be a big leap to make that possible.