Return to topic selection when dialog finishes
This matches the old behavior of MenuNPCActions. The only difference is
that dialog no longer auto-advances to the trade menu if trading is the
only available "topic". This is actually an improvement in situations
like the zombie trader in the Empyrean Campaign. Some players missed
that they could actually trade with the zombie after the dialog box
closed. Now, the dialog doesn't close, but advances to the topic
selection that shows the ability to trade.