Add support for dialog responses
In short, responses are dialog topics that are displayed alongside
a line of dialog text. Here is a simple example:

    [dialog]
    topic=New quest
    response=quest_accept
    response=quest_refuse
    him=Do you accept?

    [dialog]
    id=quest_accept
    topic=Yes
    response_only=true
    you=I accept.

    [dialog]
    id=quest_refuse
    topic=No
    response_only=true
    you=I refuse.

This will appear as:

    Do you accept?
    Yes
    No

There are three new keys in play here: id, response, and response_only.
The response key takes an ID value as its parameter. The response_only
key prevents the dialog topic from appearing in the general topic
selection list. General topics can be used as responses as long as they
have an ID to reference them with. Dialog topics don't require an ID if
you don't wish to reference them as a response.

This commit also fixes a small bug in chooseDialogNode(). The output of
npc->processDialog() wasn't being checked, so dialog with no text would
result in some breakage.